type migration

+++ src/main/java/com/kel5/ecommerce/entity/Order.java
@@ -28,7 +28,7 @@ public class Order {
     private String address;
     private String whatsapp;
     private LocalDate orderDate;
-    private float totalAmount;
+    private long totalAmount;
 
     @ManyToOne
     @JoinColumn(name = "user_id")
